+++ BLOOD_BANKING_SYSTEM-BackEnd/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "blood_banking_system-backend",
+  "version": "1.0.0",
+  "main": "server.js",
+  "scripts": {
+    "reload": "nodemon",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node server.js"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "description": "",
+  "dependencies": {
+    "body-parser": "^1.20.2",
+    "express": "^4.19.2",
+    "nodemon": "^3.1.4"
+  }
+}
