+++ BLOOD_BANKING_SYSTEM-webEnd/Pages/pride_page.html
@@ -17,5 +17,27 @@
     </ul>
   </nav>
 </header>
+<footer class="footer">
+  <div class="container">
+    <div class="footer-section">
+      <h2>About Us</h2>
+      <p>We are dedicated to providing safe and reliable blood banking services to save lives.</p>
+    </div>
+    <div class="footer-section">
+      <h2>Contact Us</h2>
+      <p>Email: contact@bloodbanking.com</p>
+      <p>Phone: +237 688827777</p>
+    </div>
+    <div class="footer-section">
+      <h2>Follow Us</h2>
+      <a href="#">Facebook</a>
+      <a href="#">Twitter</a>
+      <a href="#">Instagram</a>
+    </div>
+  </div>
+  <div class="footer-bottom">
+    <p>&copy; 2024 Blood Banking System. All rights reserved.</p>
+  </div>
+</footer>
 </body>
 </html>
