+++ package.json
@@ -9,8 +9,9 @@
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "description": "",
   "dependencies": {
     "concurrently": "^8.2.2"
-  }
+  },
+  "devDependencies": {},
+  "description": ""
 }
